Made general interface for base metrics
