+++ cpp/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(SOURCES metrics.cpp metrics.h)
+set(SOURCES metrics.cpp metrics.h metrics_utils.cpp metrics_utils.h)
 
 add_library(map_metrics SHARED ${SOURCES})
